+++ packages/icestark-app/src/AppLink.tsx
@@ -1,8 +1,24 @@
 import * as React from 'react';
 import formatUrl from './util/formatUrl';
+import removeTrailingSlash from './util/removeTrailingSlash';
+
+interface To {
+  /**
+   * A string representing the path link to
+   */
+  pathname: string;
+  /**
+   * A string representing the url search to
+   */
+  search?: string;
+  /**
+   * A string representing the url state to
+   */
+  state?: object;
+}
 
 export type AppLinkProps = {
-  to: string;
+  to: string | To;
   hashType?: boolean;
   replace?: boolean;
   message?: string;
@@ -11,7 +27,11 @@ export type AppLinkProps = {
 
 const AppLink = (props: AppLinkProps) => {
   const { to, hashType, replace, message, children, ...rest } = props;
-  const linkTo = formatUrl(to, hashType);
+
+  const _to = typeof to === 'object' ? removeTrailingSlash(to.pathname) + to.search : to;
+  const _state = typeof to === 'object' ? to.state : {};
+
+  const linkTo = formatUrl(_to, hashType);
   return (
     <a
       {...rest}
@@ -28,7 +48,7 @@ const AppLink = (props: AppLinkProps) => {
          */
         const changeState = window.history[replace ? 'replaceState' : 'pushState'].bind(window);
 
-        changeState({}, null, linkTo);
+        changeState(_state ?? {}, null, linkTo);
       }}
     >
       {children}