+++ packages/icestark-app/src/appHistory.ts
@@ -1,16 +1,16 @@
 import formatUrl from './util/formatUrl';
 
 const appHistory = {
-  push: (url: string, hashType?: boolean) => {
+  push: (url: string, hashType?: boolean, state?: object) => {
     window.history.pushState(
-      {},
+      state ?? {},
       null,
       formatUrl(url, hashType),
     );
   },
-  replace: (url: string, hashType?: boolean) => {
+  replace: (url: string, hashType?: boolean, state?: object) => {
     window.history.replaceState(
-      {},
+      state ?? {},
       null,
       formatUrl(url, hashType),
     );