+++ packages/icestark-app/src/util/removeTrailingSlash.ts
@@ -1,0 +1,5 @@
+const removeTrailingSlash = (str: string) => {
+  return str.charAt[str.length - 1] === '/' ?  str.slice(0, -1) : str;
+}
+
+export default removeTrailingSlash;